requirements-dev.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
git-filter-repo==2.34.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
requests==2.27.1
smmap==5.0.0
urllib3==1.26.9